move pages to container, add better messages
